requirements.txt
streamlit
python-dotenv
ijson
openai